requirements.txt
Flask>=3.0
pymupdf>=1.24 
reportlab>=4.1
google-genai>=0.3.0
google-generativeai>=0.7.2
